aider: Updated IniFile class to automatically set the configFile property to the path of "wslbackup.ini" in the script's directory, and updated the IniFile object creation in the Validate-BackupDir and Select-Distro functions.
# Aider chat conversation:

USER: When the class is initialized, the property configFile can be set using the following logic: ```
[IniFile]::new((Join-Path $PSScriptRoot "wslbackup.ini"))```
It will always the be same logic. The user will not need to provide a configFile value. The IniFile class will derive it with the logic I just provided.

ASSISTANT: Understood. We can modify the `IniFile` class to automatically set the `configFile` property to the path of "wslbackup.ini" in the script's directory. This will remove the need for the user to provide a `configFile` value when creating an `IniFile` object.

Here's the change:

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    IniFile([string]$configFile) {
        $this.configFile = $configFile
        $this.Read()
    }
=======
    IniFile() {
        $this.configFile = Join-Path $PSScriptRoot "wslbackup.ini"
        $this.Read()
    }
>>>>>>> updated
```

And we will also need to update the `IniFile` object creation in the `Validate-BackupDir` and `Select-Distro` functions:

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    $iniFile = [IniFile]::new((Join-Path $PSScriptRoot "wslbackup.ini"))
=======
    $iniFile = [IniFile]::new()
>>>>>>> updated
```

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    $iniFile = [IniFile]::new((Join-Path $PSScriptRoot "wslbackup.ini"))
=======
    $iniFile = [IniFile]::new()
>>>>>>> updated
```