aider: Instantiate the `IniFile` class at the beginning of the main code and make it globally accessible.
# Aider chat conversation:

USER: Let us instantiate the IniFile class at the beginning of the main code that is not in any function. It should only be instantiated once for the lifetime of the powershell execution. If it is globably accessable, then we can just reference it in any functions referencing it. If not, we can pass it as a parameter (by reference) to any function that needs it. Go.

ASSISTANT: Understood. We can instantiate the `IniFile` class at the beginning of the main code and make it globally accessible. This way, we can reference it in any function that needs it without having to pass it as a parameter.

Here's the change:

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
Clear-Host

# Usage
$backupDir = Validate-BackupDir
#Read-Host "Press Enter"

$distro = Select-Distro
$iniData = Read-Ini
Clear-Host
=======
$global:iniFile = [IniFile]::new()

Clear-Host

# Usage
$backupDir = Validate-BackupDir
#Read-Host "Press Enter"

$distro = Select-Distro
Clear-Host
>>>>>>> updated
```

And we will also need to update the `Validate-BackupDir` and `Select-Distro` functions to use the global `iniFile` object:

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    $iniFile = [IniFile]::new()
=======
    $iniFile = $global:iniFile
>>>>>>> updated
```

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    $iniFile = [IniFile]::new()
=======
    $iniFile = $global:iniFile
>>>>>>> updated
```