aider: Validate `iniFile.lastDistro` against `distroList` and set it to an empty string if it's not in the list.
# Aider chat conversation:

ASSISTANT: Yes, that makes sense. We can add a validation step to check if the current value of `iniFile.lastDistro` is in the `distroList`. If it's not, we can prompt the user to select a distro from the list.

Here are the steps to implement this:

1. In the `Select-Distro` function, after getting the `distroList`, we will check if `iniFile.lastDistro` is in `distroList`.
2. If it's not, we will set `iniFile.lastDistro` to an empty string to force the selection menu to appear.

Here is the code change:

```powershell
bkupwslubuntu.ps1
<<<<<<< HEAD
    # Get all installed distros minus any Docker instances.
    $distroList = Get-WSLDistros
    $iniFile = $global:iniFile

    # If there is only one distro, automatically select it
    if ($distroList.Count -eq 1) {
        $iniFile.lastDistro = $distroList[0]
        return
    }
=======
    # Get all installed distros minus any Docker instances.
    $distroList = Get-WSLDistros
    $iniFile = $global:iniFile

    # Validate iniFile.lastDistro against distroList
    if ($iniFile.lastDistro -notin $distroList) {
        $iniFile.lastDistro = ""
    }

    # If there is only one distro, automatically select it
    if ($distroList.Count -eq 1) {
        $iniFile.lastDistro = $distroList[0]
        return
    }
>>>>>>> updated
```

This change will ensure that the current value of `iniFile.lastDistro` is always valid. If it's not, the user will be prompted to select a valid distro from the list.